+++ backend/api/src/main/java/com/yat2/episode/collaboration/config/CollaborationAsyncConfig.java
@@ -17,17 +17,17 @@ public class CollaborationAsyncConfig {
 
     private static final String THREAD_PREFIX = "redis-append-";
 
-    private final CollaborationAsyncProperties collaborationAsyncProperties;
+    private final CollaborationAsyncProperties asyncProperties;
 
     @Bean(name = "redisExecutor")
     public Executor redisExecutor() {
         ThreadPoolTaskExecutor exec = new ThreadPoolTaskExecutor();
         exec.setThreadNamePrefix(THREAD_PREFIX);
-        exec.setCorePoolSize(collaborationAsyncProperties.corePoolSize());
-        exec.setMaxPoolSize(collaborationAsyncProperties.maxPoolSize());
-        exec.setQueueCapacity(collaborationAsyncProperties.queueCapacity());
-        exec.setKeepAliveSeconds(collaborationAsyncProperties.keepAliveSeconds());
-        exec.setAllowCoreThreadTimeOut(collaborationAsyncProperties.allowCoreThreadTimeout());
+        exec.setCorePoolSize(asyncProperties.corePoolSize());
+        exec.setMaxPoolSize(asyncProperties.maxPoolSize());
+        exec.setQueueCapacity(asyncProperties.queueCapacity());
+        exec.setKeepAliveSeconds(asyncProperties.keepAliveSeconds());
+        exec.setAllowCoreThreadTimeOut(asyncProperties.allowCoreThreadTimeout());
 
         exec.setRejectedExecutionHandler(dropAndLogError());
         exec.initialize();
@@ -38,13 +38,14 @@ public class CollaborationAsyncConfig {
     private RejectedExecutionHandler dropAndLogError() {
         AtomicLong dropped = new AtomicLong();
         AtomicLong lastLogMs = new AtomicLong(0);
+        long logIntervalMs = asyncProperties.dropLogIntervalMs();
         //TODO: Update가 drop 될 경우 Yjs Runner에게 알려 Sync 프로토콜로 복구
         return (r, executor) -> {
             long n = dropped.incrementAndGet();
-
             long now = System.currentTimeMillis();
             long prev = lastLogMs.get();
-            if (now - prev >= collaborationAsyncProperties.dropLogIntervalMs() && lastLogMs.compareAndSet(prev, now)) {
+
+            if (now - prev >= logIntervalMs && lastLogMs.compareAndSet(prev, now)) {
                 int qSize = executor.getQueue().size();
                 log.error("Redis queue full. Dropping tasks. dropped={}, queueSize={}", n, qSize);
             }
