+++ backend/api/src/test/java/com/yat2/episode/collaboration/yjs/YjsMessageRouterTest.java
@@ -17,7 +17,7 @@ import java.util.UUID;
 import java.util.concurrent.Executor;
 
 import com.yat2.episode.collaboration.SessionRegistry;
-import com.yat2.episode.collaboration.UpdateStreamStore;
+import com.yat2.episode.collaboration.redis.UpdateStreamStore;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
